Changed "Kierownica" to "ster".
A while ago, I committed a change to the translations to Polish because I realised, that many were very wrong. I am now realising, that there was one mistake I made. It turns out, "ster" is a Polish word, that I didn't know about at the time. I found out about it, while reading a book about air combat. I apologise for my previous mistake, believing that it was a mistranslation, though it really wasn't. "Ster" is a more fitting translation. It appears as if everything was alright now.